requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.2
playwright==1.47.0
requests==2.32.3
beautifulsoup4==4.12.3
httpx
redis